requirements.txt
# Python 3.11.8
# conda install -c conda-forge gdal
# matplotlib
# seaborn
# rasterio
# georasters
# openpyxl
Pyarrow
pandas
numpy
dash
plotly
geopandas
dash-bootstrap-components
flask-caching
traitlets
nbformat
gunicorn
python-dotenv